+++ Tests/ReferenceImages_64/StripePaymentSheetTests.EmbeddedPaymentMethodsViewSnapshotTests/testEmbeddedPaymentMethodsView_flatCheck@3x.png
